create doc will now navigate you to the doc-details
