+++ src/services/docsService.js
@@ -15,3 +15,13 @@ export const deleteDocById = (docId) => {
     })
     .then(fetch(`http://localhost:8088/documents/${docId}`, { method: "DELETE" }))
 }
+
+export const createDocument = (document) => {
+    return fetch(`http://localhost:8088/Documents`, {
+        method: "POST",
+        headers: {
+            "Content-Type": "application/json"
+        },
+        body: JSON.stringify(document)
+    }).then((res) => res.json())
+}
